# refactor: use b.Loop() to simplify the code and improve performance