#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use [unclog](https://github.com/informalsystems/unclog) to manage our changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code comments



These changes use b.Loop() to simplify the code and improve performance

Supported by Go Team, more info: https://go.dev/blog/testing-b-loop

More info can see https://go.dev/issue/73137.

Before:


```shell
go test -run=^$ -bench=. ./types -timeout=1h             
goos: darwin
goarch: arm64
pkg: github.com/cometbft/cometbft/types
cpu: Apple M4
BenchmarkEventBus/10Clients1Query1Event-10         	 3788924	       310.3 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/100Clients-10                    	 3767460	       316.6 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/1000Clients-10                   	   16606	     66811 ns/op	     544 B/op	       6 allocs/op
BenchmarkEventBus/10ClientsRandQueries1Event-10    	  754050	      1527 ns/op	    1312 B/op	      24 allocs/op
BenchmarkEventBus/100Clients#01-10                 	  511202	      2306 ns/op	    1952 B/op	      39 allocs/op
BenchmarkEventBus/1000Clients#01-10                	  457266	      2531 ns/op	    1952 B/op	      39 allocs/op
BenchmarkEventBus/10ClientsRandQueriesRandEvents-10         	 3649645	       327.0 ns/op	     417 B/op	       3 allocs/op
BenchmarkEventBus/100Clients#02-10                          	 2947570	       346.2 ns/op	     426 B/op	       3 allocs/op
BenchmarkEventBus/1000Clients#02-10                         	  737244	      1460 ns/op	     718 B/op	      10 allocs/op
BenchmarkEventBus/10Clients1QueryRandEvents-10              	 3672014	       322.4 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/100Clients#03-10                          	 3715051	       322.8 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/1000Clients#03-10                         	  207978	      5357 ns/op	     544 B/op	       6 allocs/op
BenchmarkMakePartSet/nParts=1-10                            	   55294	     20497 ns/op
BenchmarkMakePartSet/nParts=2-10                            	   29437	     40110 ns/op
BenchmarkMakePartSet/nParts=3-10                            	   20059	     59683 ns/op
BenchmarkMakePartSet/nParts=4-10                            	   15042	     79628 ns/op
BenchmarkMakePartSet/nParts=5-10                            	   12037	     99587 ns/op
BenchmarkProposalWriteSignBytes-10                          	 4153146	       290.1 ns/op
BenchmarkProposalSign-10                                    	   99568	     11996 ns/op
BenchmarkProposalVerifySignature-10                         	 4080648	       293.9 ns/op
BenchmarkValidatorSetCopy-10                                	   59730	     19908 ns/op
BenchmarkUpdates-10                                         	    1306	    909655 ns/op
Benchmark_2_3_Maj-10                                        	     241	   5133499 ns/op
PASS
ok  	github.com/cometbft/cometbft/types	35.253s
```



after:

```shell
go test -run=^$ -bench=. ./types -timeout=1h  
goos: darwin
goarch: arm64
pkg: github.com/cometbft/cometbft/types
cpu: Apple M4
BenchmarkEventBus/10Clients1Query1Event-10         	 3789516	       313.8 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/100Clients-10                    	 3773624	       318.5 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/1000Clients-10                   	   16605	     72004 ns/op	     544 B/op	       6 allocs/op
BenchmarkEventBus/10ClientsRandQueries1Event-10    	  673412	      1728 ns/op	    1440 B/op	      27 allocs/op
BenchmarkEventBus/100Clients#01-10                 	  501636	      2356 ns/op	    1952 B/op	      39 allocs/op
BenchmarkEventBus/1000Clients#01-10                	  450982	      2539 ns/op	    1952 B/op	      39 allocs/op
BenchmarkEventBus/10ClientsRandQueriesRandEvents-10         	 3622975	       328.3 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/100Clients#02-10                          	 3437947	       338.0 ns/op	     425 B/op	       3 allocs/op
BenchmarkEventBus/1000Clients#02-10                         	  157407	      6593 ns/op	    2052 B/op	      41 allocs/op
BenchmarkEventBus/10Clients1QueryRandEvents-10              	 3715688	       321.9 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/100Clients#03-10                          	 3714166	       322.3 ns/op	     416 B/op	       3 allocs/op
BenchmarkEventBus/1000Clients#03-10                         	  194089	      5703 ns/op	     543 B/op	       5 allocs/op
BenchmarkMakePartSet/nParts=1-10                            	   55609	     20132 ns/op
BenchmarkMakePartSet/nParts=2-10                            	   30139	     39803 ns/op
BenchmarkMakePartSet/nParts=3-10                            	   20073	     59629 ns/op
BenchmarkMakePartSet/nParts=4-10                            	   15074	     79577 ns/op
BenchmarkMakePartSet/nParts=5-10                            	   12037	     99553 ns/op
BenchmarkProposalWriteSignBytes-10                          	 4253791	       282.7 ns/op
BenchmarkProposalSign-10                                    	   99486	     12026 ns/op
BenchmarkProposalVerifySignature-10                         	 4208140	       288.5 ns/op
BenchmarkValidatorSetCopy-10                                	   59937	     19850 ns/op
BenchmarkUpdates-10                                         	    1312	    904521 ns/op
Benchmark_2_3_Maj-10                                        	     240	   4960945 ns/op
PASS
ok  	github.com/cometbft/cometbft/types	34.504s
```